refactor: move date range functions to shared library and update imports
