+++ lib/shared/package.json
@@ -2,7 +2,9 @@
   "name": "@quick-learn/shared",
   "version": "0.0.1",
   "dependencies": {
-    "tslib": "^2.3.0"
+    "tslib": "^2.3.0",
+    "date-fns": "^3.6.0",
+    "date-fns-tz": "^3.2.0"
   },
   "type": "commonjs",
   "main": "./src/index.js",
